Make TemplateControlBar its own component
