+++ ui/src/dashboards/components/Dashboard.js
@@ -1,10 +1,8 @@
 import React, {PropTypes} from 'react'
 import classnames from 'classnames'
 
-import omit from 'lodash/omit'
-
+import TemplateControlBar from 'src/dashboards/components/TemplateControlBar'
 import LayoutRenderer from 'shared/components/LayoutRenderer'
-import Dropdown from 'shared/components/Dropdown'
 
 const Dashboard = ({
   source,
@@ -48,39 +46,11 @@ const Dashboard = ({
         {'presentation-mode': inPresentationMode}
       )}
     >
-      <div className="template-control-bar">
-        <h1 className="template-control--heading">Template Variables</h1>
-        {dashboard.templates.map(({id, values, tempVar}) => {
-          const items = values.map(value => ({...value, text: value.value}))
-          const selectedItem = items.find(item => item.selected) || items[0]
-          const selectedText = selectedItem && selectedItem.text
-
-          // TODO: change Dropdown to a MultiSelectDropdown, `selected` to
-          // the full array, and [item] to all `selected` values when we update
-          // this component to support multiple values
-          return (
-            <div key={id} className="template-control--dropdown">
-              <Dropdown
-                items={items}
-                buttonSize="btn-xs"
-                selected={selectedText || 'Loading...'}
-                onChoose={item =>
-                  onSelectTemplate(id, [item].map(x => omit(x, 'text')))}
-              />
-              <label className="template-control--label">
-                {tempVar}
-              </label>
-            </div>
-          )
-        })}
-        <button
-          className="btn btn-primary btn-sm template-control--manage"
-          onClick={onOpenTemplateManager}
-        >
-          <span className="icon cog-thick" />
-          Manage
-        </button>
-      </div>
+      <TemplateControlBar
+        templates={dashboard.templates}
+        onSelectTemplate={onSelectTemplate}
+        onOpenTemplateManager={onOpenTemplateManager}
+      />
       {cells.length
         ? <LayoutRenderer
             templates={templatesIncludingDashTime}
